Aplicacion.java
This application was initially designed to help iranian people find their missing loved ones. The list of the people contains some of the information of them. By inserting the name and the surname of one person and introducing  the place and the time you have seen them plus your phone number, The contact of the person will receive  a message and will be known! Also keep in mind that your phone number will be provided to them for further investigation. When you register a new person it will be grabbed in our database and shortly after studying your petition it will added to the list.